# openapi.yaml
openapi: 3.1.0
info:
  title: WinMyDispute API
  version: 3.1.0
  description: |
    GPT-facing API for the WinMyDispute free-preview and premium-upgrade flow.
    The free tier returns a likely reason-code match, a confidence read, and exactly one strategy tip.
    Premium output unlocks after a one-time Stripe Checkout payment tied to the user's email and can produce
    a professional dispute packet as JSON plus downloadable PDF, Word-compatible RTF, or text report output.
  contact:
    name: Daniel Neville
    email: pureplays@icloud.com
    url: https://pure-plays.com
  license:
    name: Proprietary
    url: https://pure-plays.com/terms
servers:
  - url: https://api.pure-plays.com
    description: Production custom domain
  - url: https://winmydispute-api.onrender.com
    description: Render deployment
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/check-license:
    get:
      tags: [auth]
      summary: Check whether an email has premium access
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
            example: user@example.com
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
          description: Matching Stripe Checkout session ID used to exchange a paid license for a premium access token.
        - name: checkoutSessionId
          in: query
          required: false
          schema:
            type: string
          description: Alias for sessionId.
      responses:
        "200":
          description: License state for the email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Request limit exceeded for the current rate-limit window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: A valid email address is required.
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [error]
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: Too many requests
        requestId:
          type: string
          format: uuid
          nullable: true
        retryAfterSeconds:
          type: integer
          minimum: 1
          example: 42
      required: [error, requestId, retryAfterSeconds]
...
    LicenseCheckResponse:
      type: object
      properties:
        licensed:
          type: boolean
          example: false
        status:
          type: string
          example: unpaid
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
          description: Returned only when the caller provides the matching Stripe session ID for a paid license.
      required: [licensed, status, email]
...